Fixed crash with arguru synth, blitz and gamefx on 22Khz samplerate (could happen on other samplerates, but less probable). Fixed song loading problem with the mixer (Somehow it only appeared in the x64 release version, but it was a bug). Fixed wrong knob being selected on rightclick the vumeter or slider text in mixer parameters. Improvements on vst loading (prevented loading the dll twice when checking if it was a jbridge dll). Other improvements on cache generation (progress was not printing when regenerating the cache).

+++ trunk/psycle-plugins/src/psycle/plugins/arguru_synth_2_final/SynthTrack.cpp
@@ -102,12 +102,12 @@ void CSynthTrack::NoteOn(int note)
 	float nnote=(float)note+
 		(float)syntp->globalfinetune*0.00389625f+
 		(float)syntp->globaldetune;
-	OSC1Speed=(float)pow(2.0, nnote/12.0)*wavetableCorrection;
+	OSC1Speed=powf(2.f, nnote/12.f)*wavetableCorrection;
 
 	float note2=nnote+
 		syntp->osc2finetune+
 		syntp->osc2detune;
-	OSC2Speed=(float)pow(2.0, note2/12.0)*wavetableCorrection;
+	OSC2Speed=powf(2.f, note2/12.f)*wavetableCorrection;
 
 	//if (oscglide == 0.0f)
 	if (sp_cmd != 0x03)
